Add files via upload
This file allows the user to obtain the mean, stdev, median and number of reads found in a fastq file. The program will ask the user to input the directory where the .fastq files are, and will ignore any other files with a different extension. The output will be in the same directory as a .csv file called "Summary_output.csv"